backend/requirements.txt
feedparser==6.0.12
sgmllib3k==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
yfinance>=0.2.40
python-dotenv>=1.0.0
google-generativeai>=0.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
apscheduler>=3.10.4
supabase>=2.3.4
